pyproject.toml
[project]
name = "agent-cortex"
version = "0.1.0"
description = ""
authors = [
    {name = "Jacob MacInnis",email = "jacobcmacinnis82@gmail.com"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "chromadb (>=1.0.9,<2.0.0)",
    "fastapi (==0.115.9)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "transformers (>=4.52.1,<5.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "torch (>=2.7.0,<3.0.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "duckduckgo-search (>=8.0.2,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "langchain-ollama (>=0.3.3,<0.4.0)",
    "langchain-chroma (>=0.2.4,<0.3.0)",
    "langchain-huggingface (>=0.2.0,<0.3.0)",
    "langchain-openai (>=0.3.17,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-community (>=0.3.24,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
